Request unnecessary web dyno, at hobby size
Hmm ok, even with `quantity: 0`, it still tries to create a web dyno at free size, which breaks the build process.

So, fine. Let's create a web dyno, and include setup instructions to turn it off. (It's not a big deal if it stays on anyway, since it will go to sleep and not generally consume resources, and requests will just error out.)